+++ src/ml/optimization/gradient_descent-inl.hpp
@@ -187,7 +187,7 @@ inline solver_return gradient_descent(first_order_opt_interface& model,
         stats.status = OPTIMIZATION_STATUS::OPT_ITERATION_LIMIT;
       }
     }
-    stats.iters = iters;
+    stats.iters = static_cast<int>(iters);
     stats.residual = residual;
     stats.gradient = gradient;
     stats.func_value = func_value;
